fix(reservation): fix appearance after creating reservation
