+++ src/modules/main/pages/Venue/index.tsx
@@ -338,7 +338,7 @@ export default function Venue() {
                         `date=${format(date, 'yyyy/MM/dd')}`,
                         `time=${time?.join(',')}`,
                       ];
-                      navigate(`/reserve/create?${query.join('&')}`);
+                      navigate(`/reservation/create?${query.join('&')}`);
                     }}
                   >
                     確認選擇時段
